.scripts/requirements.txt
pykakasi==2.2.0
pillow-heif==1.0.0
pillow==11.3.0
python-docx==1.2.0